optional support for binary files
Added an optional argument to constructor to specify the mode argument to the file open command.  Defaults to 'r', for backwards-compatibility.
The specific use case this addresses is supporting binary-encoded encoded files by specifying the mode as 'rb'.  While using this module to parse pcap files generated by Wireshark, I found this to be a necessary change.